Fixes to addWeb3IdCredentials
